# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  ans-spareparts-api_pkg_response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  internal_features_inventory.InventoryListResponse:
    properties:
      inventories:
        items:
          $ref: '#/definitions/internal_features_inventory.InventoryResponse'
        type: array
      total:
        format: int64
        type: integer
    type: object
  internal_features_inventory.InventoryResponse:
    properties:
      id:
        type: integer
      productID:
        type: integer
      quantity:
        type: integer
    type: object
  internal_features_inventory.UpdateQuantityRequest:
    properties:
      productID:
        type: integer
      quantity:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: suppourt@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A Point of sale system API built with Clean Architechture
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0html
  termsOfService: http://swagger.io/terms/
  title: ANS Sparepart API
  version: "1.0"
paths:
...
  /inventory:
    get:
      consumes:
      - application/json
      description: Get all inventory records (admin/manager only)
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_features_inventory.InventoryListResponse'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get inventories from query
      tags:
      - inventory
    patch:
      consumes:
      - application/json
      description: Update inventory stock (admin/manager only)
      parameters:
      - description: Inventory update request
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/internal_features_inventory.UpdateQuantityRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_features_inventory.InventoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update quantity inventory
      tags:
      - inventory
  /inventory/{inventoryid}:
    get:
      consumes:
      - application/json
      description: Get inventory detail by inventory id
      parameters:
      - description: inventory ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_features_inventory.InventoryListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get inventory by ID
  /product/{productID}/inventory:
    get:
      consumes:
      - application/json
      description: Get inventory details by product ID
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_features_inventory.InventoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get inventory by product ID
      tags:
      - inventory
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization Type "Bearer" followed by a space and JWT token.
    type: apiKey
swagger: "2.0"